Version 1.1
* Fix some vanilla optics still having zoom on ironsights
* Fix vanilla optics config missing from the built addon
* Config overhaul, build script
* Minor styling fixes and improvements
