+++ source/dnct_ironsight/cup_static/config.cpp
@@ -1,4 +1,4 @@
-#include "..\defines.cpp"
+#include "..\defines.hpp"
 
 class CfgPatches
 {
